requirements.txt
discord.py>=2.3.2
firebase-admin>=6.2.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
Flask[async]>=2.3.3
Flask-CORS>=4.0.0
uvicorn>=0.22.0
gunicorn>=21.2.0
python-dateutil>=2.8.2